Living With a VP Shunt-How To Improve the Quality Of Life?
Many questions may arise in the mind of the patient before undergoing any surgery. Most people who undergo a VP shunt procedure want to know what it's like to live with a VP shunt. Here we will discuss the same with our eminent neurosurgeon. We have also mentioned the ways how to improve the quality of life while living with a VP shunt.